refactor: convert file name to PascalCase
