+++ src/components/ActionButtonGroup/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Button from "components/Commons/Buttons";
-import Image from "components/Commons/Images";
+import Button from "components/Common/Button";
+import Image from "components/Common/Image";
 import {
   ALT_TAG_ICON_CHECK,
   ALT_TAG_ICON_DELETE,
