+++ src/components/HomeContainer/index.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import Button from "components/Commons/Buttons";
+import Button from "components/Common/Button";
 import CreateTask from "components/Task/CreateNew";
 import TaskCard from "components/Task/ExistingTask/TaskCard";
 import FilterButtonGroup from "components/FilterButtonGroup";
