+++ src/pages/Home.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import HomeContainer from "components/HomeContainer";
-import NavBar from "components/Navbars";
+import NavBar from "components/Navbar";
 
 function Home() {
   return (
